move custom_logger to its own file
